+++ CHANGELOG.md
@@ -1,5 +1,10 @@
-# Pending
+# Merged
 
+* VAST Error Tracking - part 3: alerting DSP more errors [#9714](https://github.com/Applifier/unity-ads-webview/pull/9714)
+
+# Wed May 22 17:00:16 UTC 2019
+
+* Early CTA v2 experiment [#9683](https://github.com/Applifier/unity-ads-webview/pull/9683)
 * 3.1.0 Merge changes to master [#9712](https://github.com/Applifier/unity-ads-webview/pull/9712)
 
 # Tue May 21 17:16:32 UTC 2019